+++ internal/dataplane/parser/translate_utils.go
@@ -3,8 +3,8 @@ package parser
 import (
 	"fmt"
 	"reflect"
+	"strings"
 
-	"github.com/blang/semver/v4"
 	"github.com/kong/go-kong/kong"
 	"github.com/sirupsen/logrus"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -12,17 +12,11 @@ import (
 	gatewayv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
 
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane/kongstate"
+	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane/parser/translators"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/store"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/types"
-	"github.com/kong/kubernetes-ingress-controller/v2/internal/util"
+	"github.com/kong/kubernetes-ingress-controller/v2/internal/versions"
 )
-
-// kongHeaderRegexPrefix is a reserved prefix string that Kong uses to determine if it should parse a header value
-// as a regex.
-const kongHeaderRegexPrefix = "~*"
-
-// MinRegexHeaderKongVersion is the minimum Kong version that supports regex header matches.
-var MinRegexHeaderKongVersion = semver.MustParse("2.8.0")
 
 // -----------------------------------------------------------------------------
 // Translate Utilities - Gateway
@@ -40,9 +34,9 @@ func convertGatewayMatchHeadersToKongRouteMatchHeaders(headers []gatewayv1beta1.
 				string(header.Name))
 		}
 		if header.Type != nil && *header.Type == gatewayv1beta1.HeaderMatchRegularExpression {
-			if util.GetKongVersion().LT(MinRegexHeaderKongVersion) {
+			if !versions.GetKongVersion().MajorMinorOnly().GTE(versions.RegexHeaderVersionCutoff) {
 				return nil, fmt.Errorf("Kong version %s does not support HeaderMatchRegularExpression",
-					util.GetKongVersion().String())
+					versions.GetKongVersion().Full().String())
 			}
 			convertedHeaders[string(header.Name)] = []string{kongHeaderRegexPrefix + header.Value}
 		} else if header.Type == nil || *header.Type == gatewayv1beta1.HeaderMatchExact {
@@ -154,4 +148,16 @@ func generateKongServiceFromBackendRef[
 	}
 
 	return service, nil
+}
+
+// maybePrependRegexPrefix takes a path string and returns it withthe regex prefix prepended if the Kong version
+// requires it and it is not already present. If those conditions are not met, it returns the original path string.
+func maybePrependRegexPrefix(path string) string {
+	// this regex matches if the path _is not_ considered a regex by Kong 2.x
+	if LegacyRegexPathExpression.FindString(path) == "" {
+		if !strings.HasPrefix(path, translators.KongPathRegexPrefix) {
+			path = translators.KongPathRegexPrefix + path
+		}
+	}
+	return path
 }