+++ internal/dataplane/parser/translators/ingress_test.go
@@ -38,7 +38,7 @@ func TestTranslateIngress(t *testing.T) {
 						IngressRuleValue: netv1.IngressRuleValue{
 							HTTP: &netv1.HTTPIngressRuleValue{
 								Paths: []netv1.HTTPIngressPath{{
-									Path: "/api",
+									Path: "/api/",
 									Backend: netv1.IngressBackend{
 										Service: &netv1.IngressServiceBackend{
 											Name: "test-service",
@@ -75,7 +75,78 @@ func TestTranslateIngress(t *testing.T) {
 					Route: kong.Route{
 						Name:              kong.String("default.test-ingress.test-service.konghq.com.80"),
 						Hosts:             kong.StringSlice("konghq.com"),
-						Paths:             kong.StringSlice("/api$", "/api/"), // Prefix pathing is the default behavior when no pathtype is defined
+						Paths:             kong.StringSlice("/api/"), // default ImplementationSpecific
+						PreserveHost:      kong.Bool(true),
+						Protocols:         kong.StringSlice("http", "https"),
+						RegexPriority:     kong.Int(0),
+						StripPath:         kong.Bool(false),
+						ResponseBuffering: kong.Bool(true),
+						RequestBuffering:  kong.Bool(true),
+					},
+				}},
+				Backends: []kongstate.ServiceBackend{{
+					Name:      "test-service",
+					Namespace: corev1.NamespaceDefault,
+					PortDef: kongstate.PortDef{
+						Mode:   kongstate.PortModeByNumber,
+						Number: 80,
+					},
+				}},
+			}},
+		},
+
+		{
+			name: "a basic ingress resource with a single rule and prefix path type",
+			ingress: &netv1.Ingress{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test-ingress",
+					Namespace: corev1.NamespaceDefault,
+				},
+				Spec: netv1.IngressSpec{
+					Rules: []netv1.IngressRule{{
+						Host: "konghq.com",
+						IngressRuleValue: netv1.IngressRuleValue{
+							HTTP: &netv1.HTTPIngressRuleValue{
+								Paths: []netv1.HTTPIngressPath{{
+									Path:     "/api",
+									PathType: &pathTypePrefix,
+									Backend: netv1.IngressBackend{
+										Service: &netv1.IngressServiceBackend{
+											Name: "test-service",
+											Port: netv1.ServiceBackendPort{
+												Name:   "http",
+												Number: 80,
+											},
+										},
+									},
+								}},
+							},
+						},
+					}},
+				},
+			},
+			expected: []*kongstate.Service{{
+				Namespace: corev1.NamespaceDefault,
+				Service: kong.Service{
+					Name:           kong.String("default.test-ingress.test-service.80"),
+					Host:           kong.String("test-service.default.80.svc"),
+					ConnectTimeout: kong.Int(int(defaultServiceTimeout.Milliseconds())),
+					Path:           kong.String("/"),
+					Port:           kong.Int(80),
+					Protocol:       kong.String("http"),
+					Retries:        kong.Int(defaultRetries),
+					ReadTimeout:    kong.Int(int(defaultServiceTimeout.Milliseconds())),
+					WriteTimeout:   kong.Int(int(defaultServiceTimeout.Milliseconds())),
+				},
+				Routes: []kongstate.Route{{
+					Ingress: util.K8sObjectInfo{
+						Name:      "test-ingress",
+						Namespace: corev1.NamespaceDefault,
+					},
+					Route: kong.Route{
+						Name:              kong.String("default.test-ingress.test-service.konghq.com.80"),
+						Hosts:             kong.StringSlice("konghq.com"),
+						Paths:             kong.StringSlice("/api/", "/api$"),
 						PreserveHost:      kong.Bool(true),
 						Protocols:         kong.StringSlice("http", "https"),
 						RegexPriority:     kong.Int(0),
@@ -287,7 +358,7 @@ func TestTranslateIngress(t *testing.T) {
 					Route: kong.Route{
 						Name:              kong.String("default.test-ingress.test-service.konghq.com.80"),
 						Hosts:             kong.StringSlice("konghq.com"),
-						Paths:             kong.StringSlice("/v1/api$", "/v1/api/"),
+						Paths:             kong.StringSlice("/v1/api/"),
 						PreserveHost:      kong.Bool(true),
 						Protocols:         kong.StringSlice("http", "https"),
 						RegexPriority:     kong.Int(0),
@@ -493,9 +564,9 @@ func TestTranslateIngress(t *testing.T) {
 						Name:  kong.String("default.test-ingress.test-service.konghq.com.80"),
 						Hosts: kong.StringSlice("konghq.com"),
 						Paths: kong.StringSlice(
-							"/v1/api$", "/v1/api/",
-							"/v2/api$", "/v2/api/",
-							"/v3/api$", "/v3/api/",
+							"/v1/api",
+							"/v2/api",
+							"/v3/api",
 							"/other/path/1$",
 							"/other/path/2",
 							"/",
@@ -633,9 +704,9 @@ func TestTranslateIngress(t *testing.T) {
 					Route: kong.Route{
 						Name: kong.String("default.test-ingress.test-service..80"),
 						Paths: kong.StringSlice(
-							"/v1/api$", "/v1/api/",
-							"/v2/api$", "/v2/api/",
-							"/v3/api$", "/v3/api/",
+							"/v1/api",
+							"/v2/api",
+							"/v3/api",
 							"/other/path/1$",
 							"/other/path/2",
 							"/",
@@ -725,7 +796,7 @@ func TestTranslateIngress(t *testing.T) {
 							Route: kong.Route{
 								Name:              kong.String("default.test-ingress.test-service1.konghq.com.80"),
 								Hosts:             kong.StringSlice("konghq.com"),
-								Paths:             kong.StringSlice("/v1/api$", "/v1/api/"),
+								Paths:             kong.StringSlice("/v1/api"),
 								PreserveHost:      kong.Bool(true),
 								Protocols:         kong.StringSlice("http", "https"),
 								RegexPriority:     kong.Int(0),
@@ -767,7 +838,7 @@ func TestTranslateIngress(t *testing.T) {
 							Route: kong.Route{
 								Name:              kong.String("default.test-ingress.test-service2.konghq.com.80"),
 								Hosts:             kong.StringSlice("konghq.com"),
-								Paths:             kong.StringSlice("/v2/api$", "/v2/api/"),
+								Paths:             kong.StringSlice("/v2/api"),
 								PreserveHost:      kong.Bool(true),
 								Protocols:         kong.StringSlice("http", "https"),
 								RegexPriority:     kong.Int(0),
@@ -792,12 +863,12 @@ func TestTranslateIngress(t *testing.T) {
 
 	for _, tt := range tts {
 		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, TranslateIngress(tt.ingress), tt.expected)
+			assert.Equal(t, TranslateIngress(tt.ingress, false), tt.expected)
 		})
 	}
 }
 
-func Test_pathsFromIngressPaths(t *testing.T) {
+func Test_PathsFromIngressPaths(t *testing.T) {
 	for _, tt := range []struct {
 		name string
 		in   netv1.HTTPIngressPath
@@ -810,8 +881,8 @@ func Test_pathsFromIngressPaths(t *testing.T) {
 				PathType: &pathTypePrefix,
 			},
 			out: kong.StringSlice(
-				"/v1/api/packages$",
 				"/v1/api/packages/",
+				"/v1/api/packages$",
 			),
 		},
 		{
@@ -821,8 +892,8 @@ func Test_pathsFromIngressPaths(t *testing.T) {
 				PathType: &pathTypePrefix,
 			},
 			out: kong.StringSlice(
-				"/v1/api/packages$",
 				"/v1/api/packages/",
+				"/v1/api/packages$",
 			),
 		},
 		{
@@ -855,7 +926,7 @@ func Test_pathsFromIngressPaths(t *testing.T) {
 				Path:     "",
 				PathType: &pathTypeExact,
 			},
-			out: kong.StringSlice("/"),
+			out: kong.StringSlice("/$"),
 		},
 		{
 			name: "path type implementation-specific will leave the path alone",
@@ -867,7 +938,8 @@ func Test_pathsFromIngressPaths(t *testing.T) {
 		},
 	} {
 		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.out, pathsFromIngressPaths(tt.in))
+			// TODO split test cases to handle regex
+			assert.Equal(t, tt.out, PathsFromIngressPaths(tt.in, false))
 		})
 	}
 }
@@ -926,6 +998,166 @@ func Test_flattenMultipleSlashes(t *testing.T) {
 	} {
 		t.Run(tt.name, func(t *testing.T) {
 			assert.Equal(t, tt.out, flattenMultipleSlashes(tt.in))
+		})
+	}
+}
+
+func TestTranslateIngressRegexPrefix(t *testing.T) {
+	tts := []struct {
+		name     string
+		ingress  *netv1.Ingress
+		expected []*kongstate.Service
+	}{
+		{
+			name: "a basic ingress resource with a single rule and prefix path type",
+			ingress: &netv1.Ingress{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test-ingress",
+					Namespace: corev1.NamespaceDefault,
+				},
+				Spec: netv1.IngressSpec{
+					Rules: []netv1.IngressRule{{
+						Host: "konghq.com",
+						IngressRuleValue: netv1.IngressRuleValue{
+							HTTP: &netv1.HTTPIngressRuleValue{
+								Paths: []netv1.HTTPIngressPath{{
+									Path:     "/api",
+									PathType: &pathTypePrefix,
+									Backend: netv1.IngressBackend{
+										Service: &netv1.IngressServiceBackend{
+											Name: "test-service",
+											Port: netv1.ServiceBackendPort{
+												Name:   "http",
+												Number: 80,
+											},
+										},
+									},
+								}},
+							},
+						},
+					}},
+				},
+			},
+			expected: []*kongstate.Service{{
+				Namespace: corev1.NamespaceDefault,
+				Service: kong.Service{
+					Name:           kong.String("default.test-ingress.test-service.80"),
+					Host:           kong.String("test-service.default.80.svc"),
+					ConnectTimeout: kong.Int(int(defaultServiceTimeout.Milliseconds())),
+					Path:           kong.String("/"),
+					Port:           kong.Int(80),
+					Protocol:       kong.String("http"),
+					Retries:        kong.Int(defaultRetries),
+					ReadTimeout:    kong.Int(int(defaultServiceTimeout.Milliseconds())),
+					WriteTimeout:   kong.Int(int(defaultServiceTimeout.Milliseconds())),
+				},
+				Routes: []kongstate.Route{{
+					Ingress: util.K8sObjectInfo{
+						Name:      "test-ingress",
+						Namespace: corev1.NamespaceDefault,
+					},
+					Route: kong.Route{
+						Name:              kong.String("default.test-ingress.test-service.konghq.com.80"),
+						Hosts:             kong.StringSlice("konghq.com"),
+						Paths:             kong.StringSlice("/api/", "~/api$"),
+						PreserveHost:      kong.Bool(true),
+						Protocols:         kong.StringSlice("http", "https"),
+						RegexPriority:     kong.Int(0),
+						StripPath:         kong.Bool(false),
+						ResponseBuffering: kong.Bool(true),
+						RequestBuffering:  kong.Bool(true),
+					},
+				}},
+				Backends: []kongstate.ServiceBackend{{
+					Name:      "test-service",
+					Namespace: corev1.NamespaceDefault,
+					PortDef: kongstate.PortDef{
+						Mode:   kongstate.PortModeByNumber,
+						Number: 80,
+					},
+				}},
+			}},
+		},
+	}
+
+	for _, tt := range tts {
+		t.Run(tt.name, func(t *testing.T) {
+			assert.Equal(t, TranslateIngress(tt.ingress, true), tt.expected)
+		})
+	}
+}
+
+func Test_PathsFromIngressPathsRegexPrefix(t *testing.T) {
+	for _, tt := range []struct {
+		name string
+		in   netv1.HTTPIngressPath
+		out  []*string
+	}{
+		{
+			name: "path type prefix will expand the match to a trailing slash if not provided",
+			in: netv1.HTTPIngressPath{
+				Path:     "/v1/api/packages",
+				PathType: &pathTypePrefix,
+			},
+			out: kong.StringSlice(
+				"/v1/api/packages/",
+				"~/v1/api/packages$",
+			),
+		},
+		{
+			name: "path type prefix will expand the match with a literal match if a slash is provided",
+			in: netv1.HTTPIngressPath{
+				Path:     "/v1/api/packages/",
+				PathType: &pathTypePrefix,
+			},
+			out: kong.StringSlice(
+				"/v1/api/packages/",
+				"~/v1/api/packages$",
+			),
+		},
+		{
+			name: "path type prefix will provide a default when no path is provided",
+			in: netv1.HTTPIngressPath{
+				Path:     "",
+				PathType: &pathTypePrefix,
+			},
+			out: kong.StringSlice("/"),
+		},
+		{
+			name: "path type exact will cause an exact matching path on a regular path",
+			in: netv1.HTTPIngressPath{
+				Path:     "/v1/api/packages",
+				PathType: &pathTypeExact,
+			},
+			out: kong.StringSlice("~/v1/api/packages$"),
+		},
+		{
+			name: "path type exact will cause an exact matching path on a regular path with a / suffix",
+			in: netv1.HTTPIngressPath{
+				Path:     "/v1/api/packages/",
+				PathType: &pathTypeExact,
+			},
+			out: kong.StringSlice("~/v1/api/packages/$"),
+		},
+		{
+			name: "path type exact will supply a default if no path is provided",
+			in: netv1.HTTPIngressPath{
+				Path:     "",
+				PathType: &pathTypeExact,
+			},
+			out: kong.StringSlice("~/$"),
+		},
+		{
+			name: "path type implementation-specific will leave the path alone",
+			in: netv1.HTTPIngressPath{
+				Path:     "/asdfasd9jhf09432$",
+				PathType: &pathTypeImplementationSpecific,
+			},
+			out: kong.StringSlice("/asdfasd9jhf09432$"),
+		},
+	} {
+		t.Run(tt.name, func(t *testing.T) {
+			assert.Equal(t, tt.out, PathsFromIngressPaths(tt.in, true))
 		})
 	}
 }